requirements.txt
gradio
moviepy==1.0.3
opencv-python
yt-dlp
openai-whisper
google-generativeai
openai
anthropic
deepseek
openrouter
webrtcvad-wheels
python-dotenv
textwrap3
Pillow
protobuf<5
requests
supervision
tqdm
numpy
